Moved country grid page to pages folder
